Adding -z: flags
Never again shall we say that Zeromus is too big and/or too bad!
Because *now* Zeromus can be absolutely busted. Big thanks to Wylem
for help with bugfixes and understanding parts of the vanilla code;
half these flags wouldn't exist/be as good without his input.

Here's a rundown of the changes:

-z:physical: Zeromus usually has a magic-based script.
Using -z:physical, all Big Bangs (and the one Meteo cast) are now
Dark Waves, using a buffed physical attack (two attack indices have
been repurposed) to roughly match the damage that Big Bangs do.
Zeromus has replaced Virus with Needle/Counter, the counter Nuke with
a very strong punch, and the scripted Nuke with... something special.
The counter scripts are entirely different: instead of reacting to
magic, Zeromus now reacts to Fight (with Fight), Dart (with Counter),
Aim (with target-all Counter), and Jump (with a buffed Fight). All of
the thresholds are all still the same, though.

-z:physmag: 50% of the time, nothing happens, and 50% of the time it's
as if you set -z:physical before rolling the seed.

-z:whichbang: A popular request! Now every Big Bang is randomized to
a spell chosen from a short list, some more dangerous than others.
Expect to reset on the Z fight because you forgot to Wall up and then
got blasted by MegaNuke.

-z:nonerfs: Was it the devs' intention to let you decrease the power
of Big Bangs by triggering counter-Nukes? Probably, given PR, but
this flag puts chains around every Big Bang to prevent you from nerfing
them. Every Big Bang will now be full power (8 or 9 rolls depending on
which one). As a note: when this flag is combined with -z:whichbang,
some spells don't get as many rolls because 9 rolls of Meteo is pretty
unnecessary if you can't nerf it. (... but now you might *not* want
to nerf it!)

-z:mustnerf: Sometimes, if you have enough HP, it just... doesn't
matter if you don't nerf the Big Bangs. Tank them, cast Cure4, carry
on. Not anymore! Now every Big Bang (or replacement attack/spell) is
strong enough to do 9999 damage (unless the formula prevents that, if
Z is casting something like an HP-based spell). To live, you must nerf
it or otherwise dodge the attack. Obviously mutually exclusive with
-z:nonerfs.

-z:phaseshift: It might be fun to start the battle, see one Big Bang,
and then immediately go into the Meteo phase. This flag just shuffles
the order in which Zeromus goes through the Virus/Nuke/Meteo attack
phases, without changing the HP thresholds or the counters (so in
the middle/last phases, Z still counters summons with Virus regardless
of what the script is doing).

-z:vanillasprite: The -vanilla:z flag has been renamed but otherwise
nothing changes about it (the Z sprite code has not been moved, though
it possibly could in the future).

Implementation notes: the zeromus_rando module has been added and is
called near when wyvern_rando/etc. are called. There's an f4c file
that contains all of the scripting changes; in that file there are
changes to Zeromus's base stats if necessary and in the case of the
physical script, there are bugfixes to avoid softlocks with redirected
spells (related to the "something special" above) and damage overflows
due to Dark Wave not capping damage at 9999. The script changes are
built in the zeromus_rando module using lists of strings; it's kind of
awkward, but it works/is mildly flexible. Specs have been compiled/etc.