Removing SomaDrops from boxes on "3 Point" wacky
It turns out that in "vanilla" FE, you can get SomaDrops in boxes
while playing the 3 Point Challenge wacky (which normally limits
your MP to 3 maximum). Since SomaDrops are removed from shops on
this wacky, I figure it's an oversight that they aren't removed
from boxes too. The fix is a quick addition to treasure_rando.py,
to refine the items_dbview to remove SomaDrops.

Note: if you are merging into a multi-wacky fork, you'll need to use
the correct syntax for checking the wacky, i.e.
    if '3point' in env.meta.get('wacky_challenge', [])