Adding -z:chaos and -z:lavosshell
These flags are answers to the questions:
1. What if Z had mostly random scripting (but still "felt" like Z)?
2. What if Z had three random monster scripts?
The goal is to provide more options for making Zeromus feel a bit more
non-trivial, or at least more interesting.

zeromus_rando.py had a large makeover. The old zeromus_scriptchange.f4c
had the Jump and Dark Wave specific parts moved out of it and into
their own f4c's, and then was also renamed. The two new flags call
their own f4c's; -z:chaos builds entirely new scripts for both actions
and reactions, while -z:lavosshell has code for allowing Zeromus to
load both alternate and normal AI scripts. uispec has been updated to
include a "no-flag" option for scripting (via <@vanilla>), to better
parcel the subflags.

Some flag-specific implementation details:
-z:chaos builds attack phases of 2-5 attacks for phases 1 and 2, and
1-3 attacks for phase 3. There is a chance to Black Hole after each
attack (up to twice), a chance for the attack to be a "shake" move
(up to twice) like vanilla Big Bang (which you can try to nerf), and
a chance to pass a turn. Every reaction now nerfs Big Bangs (instead
of only some of them). Some attacks are... very strong. Watch out.
In general, though, the attacks are curated to be mostly equivalent
power to what we see in vanilla. There should be no issue with the
60-byte-limit on AI scripts.

-z:lavosshell points Z's attack scripts at three random scripts,
where the scripts are slightly curated so as to not be completely
broken. Since we don't want condition flags to change, that restricts
the choice of script as well. The list of scripts is likely to change
in the future, because "Tornado every turn" is probably unreasonable.
Reactions are left unchanged.