Modifying Truth in Advertising (Odin)
One suggestion about Odin's weakness changes was to add an extra
empty turn during the first sword-raised phase, to allow for more
time available to hit the weakness. Odin can be kind of awful without
being able to hit the weakness for the full time, so we'll do that.
(But only during the normal script; if Bwoahdin is on, then you
have plenty of time to hit the weakness before Odin hits you with
the party-wise attack.)