Fixing a small bug in -fusoya:location
The way I implemented -fusoya:location involved adding two "fake"
slots for starting spells, so that Fu starts with 2x3 = 6 spells.
However, the check I made for "is this spell already in this slot?"
did not account for a spell being in the other starting slot, so
Fu could start with e.g. two copies of Cure1. The game doesn't
prevent this because the spells are directly written into FuSoYa's
initial data in the ROM, which is pulled into RAM when Fu is
initialized (unlike when characters have spells added to their
spell lists).

The fix is to add a secondary check that prevents exactly this
situation.