Fixing/Updating Rosa's Mom hints
The Rosa's Mom hints have needed a bit of care for a while now, so
let's fix up some stuff, both vanilla FE and fork-related.

First, the vanilla FE fix: on Owin:crystal Kvanilla, Rosa's Mom will
shout "X" at you because the code in core_rando.py that specifies
which slots have key items includes the fixed_objective slot (because
it's got a key item in it). The fix is just to remove that slot when
Owin:crystal is on (done in core_rando.py).

Next, the fork fixes. sgrunt added Kforge and Kpink, which allow
the Forge slot and the Pink Tail slot to reward you with key items;
thus the hints need to be available for those slots when appropriate.
The original code was bugged to always add those checks (and in fact,
add them a second time if the flags were on). This bug impacted:
- the hints that Rosa's Mom would give; she would always try to give
you hints for the Forge item and the Pink Tail turn-in, even when the
slots are not available. Thus she would tell you that Kokkol can forge
the Crystal (since we never changed the hint text from the very
original "Forge the Crystal" FE update) and then yell "X" at you, since
we never updated the text to give a proper hint.
- the -exp:kicheckbonus_num flag, because the number of available key
item checks depended on the length of the potential_key_item_slots
list, which was always two larger than it was supposed to be. Thus you
would get less experience than you should have.

sgrunt also added Knofree:dwarf, which moves the key item from Bedward
to Cid in Dwarf Castle. The hint text never showed up, though, because
the relevant hint text was numbered $7E and on Galeswift's branch the
$5E reward slot (-> $7E message) was for the gated objective item,
which is correctly never added to the potential_key_item_slots.

The fix is in two places; in core_rando.py, instead of adding the slots
we remove them from list if the relevant flags are not on. We also do
not add them (again) on hidden flags. Moreover, in rosas_mom_hints.f4c
we make amendments to the hint text:
- the Forge item text is now "can forge a reward when" instead of
"can forge the [crystal]Crystal when".
- the Pink Tail item text is now "A man in the Adamant Grotto dreams
of the legendary [tail]Pink.", to both actually give it a hint and
then give it flavour/make it distinct from the Rat Tail check.
- the Knofree:dwarf check, where you get an item from Cid in Dwarf
Castle, is now renumbered to be message $7F instead of $7E, to make
it show up.

Lastly, in rewards.f4c, Galeswift had changed Rewards__MaxSlotCount
from $60 to $70 while testing out putting characters in boxes; later,
the number had not gone back to $60 when those testing changes were
undone (even though the related comment immediately above it had).
This change caused Rosa's Mom to spit out random textboxes until the
loop started over. The fix is that in rewards.f4c we make it $60 again.