Softlock fix for -z:chaos
The f4c compiler does not handle 'wait' at the end of a script well;
no scripts end in $FE $FF, so the compiler tries to smush the last two blocks together and outputs a broken script. The fix for -z:chaos is to avoid putting 'wait' at the end of a script (which could happen due to the Black Hole scripting with the message).