Bugfixes for Csuperhero, Cpaladin tidying-up
What started as simply making sure your Superhero was actually powered
up and down correctly turned into dealing with Cecil being a fussy
class-changer. So, here we are!

Csuperhero: In eventextensions_misc.f4c, the code was originally just
powering up your middle-slot character (which works at the beginning
of the game but does not work at Zot; the Zot event was added later,
which explains the omission). There is now a loop to find the superhero
in your party. In rewards.f4c, the code was likely copy-pasted, so it
also just checked the middle slot; the same loop has been placed there
as well.

Then, because Cecil changes class at Ordeals, if he is your hero then
he needs to get powered up if DKC had been powered up. If Cecil is the
superhero, *and* if he doesn't start as a Paladin (see Cpaladin),
then there is now a check for the Earth Crystal right after
Cecil gets Paladinized: if you don't have the Earth Crystal, then he
gets powered up (luckily, the code is the same If conditional as the
Zot event). This change happens in mtordeals.f4c, but is triggered
in character_rando.py. Obviously if Cecil does not change class, then
he does not need to be powered up again, so that code gets subbed out.

Moreover, the MakeSuperhero event was not happening under the
Quickstart test setting, so that's been added as a substitution when
appropriate. That happens in opening.f4c.

Lastly, there was a small bug in the MakeSuperhero event; the carry
flag needs to get cleared before adding, because otherwise you might
end up with 1 extra Str point.

Cpaladin: Under this flag, Cecil becomes a Paladin between the time
you hit New Game and the opening cutscene starting. However, Cecil's
sprites and portraits are visible on the pregame screens (on the main
screen if he is the starting character, and on the two naming screens).
So, under this flag we should instead load the Paladin graphics and
palettes. These changes happen in...
- paladin_start_sprite.f4c, which hooks into the menu code for loading
sprite classes and replaces $00 (DKCecil) with $09 (PCecil).
- character_rando.py, which loads paladin_start_sprite as required.
- standing_characters_paladin.bin, a binary patch file that is the
same as the usual standing_characters.bin file except that PCecil's
standing sprite is where DKCecil's standing sprite is.
- generator.py, which loads standing_characters_paladin.bin as
required.
- pregame_screen.f4c, which has two places where we sub in PCecil's
job class $09 for DKCecil's $00, to handle 1. the palette for the
standing sprite, and 2. the portrait/palette in the naming screen.