Fix the Csuperhero/quickstart options
The quickstart option code in opening.f4c is
actually always there; it's just that if the
option isn't on, then script preprocessing subs
it out entirely, to let the usual FE opening
events happen.

Because of the way the substitution processing
is written, it's not possible to (easily) do a
substitution within a substitution, so instead I
wrote another substitution to make sure your
superhero gets powered up only under quickstart.
Unfortunately, I didn't remember to toggle that
substitution on when quickstart isn't on, so...
every starting character is now super-powered!
Great.

The fix is obviously just to add in the other
substitution, which happens in generator.py
(around line 913).