Fixing KeyError thrown in fusoya_rando
On -fusoya:location,maybe without J spells, the J spells get popped
off the dictionary, and *then* every spell (including the J spells)
have a 15% chance to be popped off the dictionary... but of course,
pop() does not appreciate it when the key you're trying to remove
does not exist. So, make it pop(..., None) to avoid throwing an
error that halts the program execution. Whoops!