Fixing Kick/Dark Wave miss issues
Under -z:physical/mustnerf, Zeromus uses Dark Wave for 9999 damage
unless you nerf it. Since the damage algorithm for Kick and Dark Wave
doesn't actually cap damage in vanilla (they just knew that you
would never normally be able to do more than 9999 damage), Zeromus
was doing "Miss!" damage or fake-healing you for huge amounts of
damage... and then you die anyway. Because of that, we patched the
lines where damage was being stored to instead cap damage at 9999.

Unfortunately, the fact that both skills can miss and have the damage
value be replaced with 0x4000, i.e. setting the flag that tells the
game to display "Miss!", means that there is a case where the damage
is a number larger than 9999 but isn't actual damage. Hence Karate can
Kick your strong-defense party members for 9999. Whoops.

The fix is to change how we patch the algorithm in
dark_wave_damage.f4c, primarily by mimicking the way the vanilla game
handles misses but with a damage cap added in afterwards. The location
that gets patched is a bit earlier than previously, meaning that we
handle any misses and the damage cap and then return to the original
code where it's ready to store damage values. There are actually two
jumps to external code, one each where normally there would be a check
for a miss (the first is Kick-specific, since the subtraction means
that you could end up with less than 0 damage, and the second is
for both skills ending up with exactly 0 damage).